### 2022-07-08 #4557 @danielskatz
👋 @phpuech - thank you for your submission.  I see a number of small errors in the paper that could be fixed, .e.g.
* there should be no space before the ";" in the affiliations
* there's an extra ")" in the acknowledgements
* the English isn't quite idiomatic in some places
*
All of these can be fixed later, but if you want to address any of them now, please feel free to make changes to your .md file, then use the command `@editorialbot generate pdf` to make a new PDF. editorialbot commands need to be the first entry in a new comment.
