### 2022-07-08 #4557 @danielskatz
In addition, I am a little confused by the software - Is it used "live" in an experimental setting, or is it data analysis software that is used more separately from an experiment (either after or before)?  I'm sorry, but this is a field I don't know well.
