### 2022-07-08 #4557 @phpuech
Hi
thanks a lot for your comments, I will check them carefully and have the MS reread by a native speaker.
The software is for data analysis and used after the experiment for processing the force curves we get from a dedicated optical tweezers set up. It allows us to extract relevant quantitative and qualitative information from the force vs. time traces, and produces a table as output, containing all the measurements, that we process, plot, sort, analyse with python codes that are not included in the present manuscript. We plan to add them as a future extension.
Sincerely
PH Puech
