### 2022-08-19 #4557 @phpuech
Hi,
I just wanted to check if everything was fine.My student wanted to know if/when we should get a feedback from JOSS.
Sincerely
PH Puech
