### 2022-09-01 #4557 @phpuech
Hi @prashjha,
thanks for your message, I was wondering I we did a mistake somewhere in the process, since it is our first submission ever to JOSS of our entire lab...
We sill go through the list and post here the names asap (should we use the @name here ?)
best
p
