### 2022-10-22 #4557 @smsaladi
Hi -

Thanks for thinking of me. Unfortunately, I don’t have capacity to review
this manuscript  right now. Sorry about that

Shyam

On Sat, Oct 22, 2022 at 8:59 AM Prashant K. Jha ***@***.***>
wrote:

> Howdy, @florian-huber <https://github.com/florian-huber>; we have this
> interesting JOSS submission related to your interests. Would you, by any
> chance, have time and are interested in reviewing this submission? Please
> let me know!
>
