### 2022-10-22 #4557 @ryEllison
Hello,

Yes. I will happily review it! I’ve been added as a reviewer already, but the review process hasn’t been initiated yet.

R

Sent from my iPhone

> On Oct 21, 2022, at 11:44 AM, ph puech ***@***.***> wrote:
>
> ﻿
> Hi,
> I just wanted to check if you had the possibility to review our work ?
> Sincerely
> PH Puech, for the authors
>
