### 2022-10-23 #4557 @prashjha
Hi @florian-huber, reviewing for JOSS requires reviewing both software and manuscript; please see [this link](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) for the review process. For your benefit, here is the main part from that link:

"A JOSS review involves checking submissions against a checklist of essential software features and details in the submitted paper. This should be objective, not subjective; it should be based on the materials in the submission as perceived without distortion by personal feelings, prejudices, or interpretations.

We encourage reviewers to file issues against the submitted repository’s issue tracker. When you have completed your review, please leave a comment in the review issue saying so.

You can include in your review links to any new issues that you the reviewer believe to be impeding the acceptance of the repository. (Similarly, if the submitted repository is a GitHub repository, mentioning the review issue URL in the submitted repository’s issue tracker will create a mention in the review issue’s history.)"
