### 2022-10-24 #4877 @editorialbot
[REVIEW]: OT-Analysis : a software for rich analysis of force curves when   probing living cells with optical tweezers
**Submitting author:** @phpuech (<a href="http://orcid.org/0000-0002-8521-0685">pierre-henri puech</a>)
**Repository:** https://github.com/phpuech/OT_Analysis
**Branch with paper.md** (empty if default branch):
**Version:** 0.2.1
