### 2022-11-23 #4877 @florian-huber
- Article text and Readme need some text edits. A first round of suggestions here: https://github.com/phpuech/OT_Analysis/pull/1
- Similar thing needs to be done for the Readme.
