### 2022-11-23 ca53f32
Create ci_pipeline.yml
