+++ JOSS/paper.tex
@@ -1,145 +1,0 @@
-\documentclass[a4paper]{article}
-\usepackage{amsmath,amssymb}
-\usepackage{lmodern}
-\usepackage{graphicx}
-%\usepackage[left=3cm,right=2cm,top=2cm,bottom=2cm]{geometry}
-
-\title{OT_Analysis : a software for rich analysis of force
-	curves when probing living cells with optical tweezers }
-\author{Thierry Galliano$^1$, Guillaume Gay$^2$,\\ Laurent Limozin$^1$, Pierre-Henri Puech$^1$}
-\date{}
-
-
-
-\begin{document}
-\maketitle
-
-\noindent $^1$ : Inserm UMR 1067 / CNRS UMR 7333 / AMU UM61 / CENTURI  ; Lab. Adhésion et Inflammation ; Marseille - France
-
-\noindent $^2$ : France Bio Imaging ; Montpellier - France
-
-\section{Summary}\label{summary}
-
-Optical tweezers are a light-based technique for micromanipulating
-objects. It allows to move objects such as microbeads and cells, and to record minute forces down to a few pN, which makes it a technique very well adapted to
-mechanical measurements on living cells \cite{gennerich_optical_2017}.
-We are interested in the mechano-transduction properties of lymphocytes. We seek to dissect the effect of forces and cell mechanics on the
-cellular response, in the context of the immune system. T cell mechanotransduction has been
-recently demonstrated to be instrumental in the finesse and accuracy of the
-response of the latter \cite{puech_mechanotransduction_2021}. Aside,
-cells can exert forces when performing their action, eg. cytotoxic T
-cells are using forces to kill target infected cells
-\cite{basu_cytotoxic_2016}.
-
-Using optical tweezers and specifically decorated beads as handles, we
-pull membrane nano-tubes from gently adhered living lymphocytes
-\cite{sadoun_dissecting_2020}. Such nanotubes are usually used to
-probe the tension of adherent cells \cite{diz-munoz_control_2010}. By
-varying the antibodies that are used to decorate the beads, we select
-the molecule type we specifically pull on, and we then explore the
-molecules which are characteristic of the immune synapse, that is one of
-the key organisational structures that has profound implications in T
-cell recognition and action \cite{baldari_immune_2017}.
-
-Using this approach, we probe not only the forces of recognition of the
-given antibody to its target molecule, but also, by using strong
-extracellular bridges, we probe the cytosolic link of the probed molecule
-to the cytoskeleton. Such a link has been proposed to be instrumental in
-the way T cells can apply or feel forces through the molecule. A
-theoretical model has been built and will be reported in a dedicated
-article \cite{manca_membrane_2022}. Aside, we will demonstrate the application of the software on full data.
-
-\begin{figure}[h]
-	\centering
-	\includegraphics[width=\linewidth]{dessin2}
-	\caption{Schematics of the experiment. A : Sequence of events (a) antibody decorated bead,  trapped by the focused infrared laser beam, and lymphocyte are far from each other, and approached ; (b) contact is formed under a controlled level of force and maintained for a given time ; (c) the cell is displaced away from the bead, eventually resulting in the pulling of a tube (d) and its rupture (e), back to the situation of (a). B : Resulting Force vs.Time curve with the indicated events. Adapted from \cite{manca_membrane_2022}}
-	\label{fig:dessin}
-\end{figure}
-
-The experimentally obtained data consists in force signal as a function of time
-(among other parameters), in the three directions of space, obtained in
-large quantities (at least 10 per cell / bead couple, and up to 20
-couples tested per sample), containing rich and detailled features that
-can relate to molecular and/or cellular mechanics that our model
[…]
