+++ README.md
@@ -119,55 +119,58 @@ Note that due to the curvature of the cells we used (T cells, 10µm in diameter)
 
 - Factor optical effect (x STD): Number of times the standard deviation to correct the optical effect. 
 
-To ease the reuse of previously used analysis parameters, a method saving/loading buttonia avalaible.
+To ease the reuse of previously used analysis parameters, method saving/loading buttons are avalaible.
 
 ## Menu after launching the analysis
 
-There are three options:
+There are three options to run the analysis
 
-- Supervised: Allows you to switch to a new window with the display of curves and a supervisory menu
-- Unsupervised: Allows you to retrieve the output file of the automatic analysis
-- ...with graphs: Allows you to retrieve the output file of the automatic analysis completed with all the graphs
+- Supervised: The computer pre-analyses the data, presents the analysis on graphs and offers to the user a selection of tools to correct, reanalyse, requalify the data. This will happen in a new dedicated window.
+- Unsupervised: The computer does it all alone, and you retrieve the output file of the automatic analysis. You cannot act on the analysis.
+- ...with graphs: Same as above, but output the graphs of the analysed curves for further consultation
 
-If we choose supervised:
+If one chooses supervised, it will open a new window with : 
 
 ### Graphic display window with supervision
 
-Visualization of all curves as a function of time on the 3 axes and as a function of distance on the main_axis
+This allows the visualization of all curves as a function of time on the 3 axes and as a function of distance on the main_axis. Time can be switched to distance.
 
 ### Supervision menu
 
-- Close supervision Panel: Possibility to close this menu for a more important visualization of the curves
+This menu contains a series of buttons and options that are briefly described below.
+
+- Close supervision Panel: Closes the menu for a larger space for visualization of the curves
 - Buttons to navigate between the curves. Can be operated with the left/right arrow keys
 - Button to save the output file with an indication that the supervision is stopped at this curve (treat_supervised column)
-- curve name
-- button for zooming with characteristic point and fit on distance curves (Pull and Press segment only)
-- if curve misaligned warning of misalignment axis with a possibility to change this status
-- fit validation of the Press segment
-- management of the optical correction with a post-analysis control
-- fit validation of the Pull segment
-- correction of the type of curve classification (type defined checked)
+- Curve name is presented
+- Zooming for setting characteristic point and fit on distance curves (Pull and Press segment only)
+- Curve misaligned warning of misalignment axis, with the possibility to change this status
+- Indication of the validity of the fit of the Press segment
+- Management of the optical correction with a post-analysis control
+- Indication of the validity of the fit of the Pull segment
+- Type of curve obtained by automatic classification (type defined checked) that can be hand modified by the user
 - Pagination to determine our position in the whole analysis. Possibility to move with the 'Enter' key and the number of the curve
 
-### Changes in points and fits
+### Changes in characteristic points and zones of fits
 
-In the supervision interface, you can modify the characteristic points and curve fits:
-Go to the force vs distance curve
-In the menu press Edit/Pick event
-in the secondary window that appears choose what you want to modify and click on OK
-then click on the graph:
-- on the selected point
-- on the two extreme points for the fit
+In the supervision interface, one can modify the characteristic points and curve fits:
+- Go to the force vs distance curve
+- In the menu press Edit/Pick event
+- in the secondary window that appears choose what you want to modify and click on OK
+- then click on the graph, either:
+    - on the selected point
+    - on the two extreme points for the fit
 
 ### Summary plot
 
-On the last curve, a yellow button appears at the bottom of the supervision table. This button displays a graphic summary window for the validation of the input data as well as the classification with respect to the thresholds.
[…]
