+++ README.md
@@ -189,7 +189,9 @@ firefox https://phpuech.github.io/user_doc.html
 ## Output file format
 
 The output data is contained in a `csv` file with 48 columns and 1 line per analyzed curve.
-Bellow is the rapid description of each column. Note that this description will soon be made more clear for a non-familiar user.
+Bellow is the rapid description of each column ; content should appear to the user trivial in the choices of naming. 
+
+Note : this description will soon be made more clear for a non-familiar user.
 
 
 ### Important data from the analysis for post-processing
@@ -197,79 +199,99 @@ Bellow is the rapid description of each column. Note that this description will 
 - treat_supervised type=bool
     True if curve is visualized otherwise False
 - automatic_type type=str
-    type determined by the automatic analysis
+    type of curve event determined by the automatic analysis (NAd, AD, TU for non adhesive, adhesive, tube resp.)
 - type type=str
-    type given to the curve with the supervisor menu. If there is no supervision then the same as the 'automatic_type' column.
+    modified type given to the curve using the supervision menu. If there is no supervision then the value is the same as the 'automatic_type' column.
 - automatic_AL type=str
-    "No" if the curve is misaligned according to the automatic threshold otherwise "Yes"
+    x,y,z alignement of the curves. The value is "No" if the curve is misaligned according to the automatic threshold.
 - AL: str
-    Readjustment through supervision. If no supervision, then same as "automatic_AL"
+    x,y,z alignement of the curves : readjustment through supervision. If no supervision, then the value is the same as "automatic_AL"
 - automatic_AL_axe type=list
-    secondary axis affected by curve misalignment and its sign to know the direction of the misalignment with respect to the direction of the main axis
+    secondary axis along which misalignment occurs, and its sign to know the direction of the misalignment with respect to the direction of the main axis (see below for the conventions for the signs)
 - optical_state type=str
-    optical correction applied (No_correction, Auto_correction, Manual_correction)
+    status of the correction for the optical artifact (No_correction, Auto_correction, Manual_correction)
 
 
-### Data of the analysis parameters
+### Analysis parameters saved for data post processing
+
+The following are setting some of the parameters for the analysis
 
 - model type=str
     model for the fit on "Press" segment chosen by the user for the analysis
-- Date type=str
-    date of creation of the curve file
-- Hour type=str
-    time of creation of the curve file
-- condition type=str
-    condition applied to the analysis set (often antibodies on the bead)
-- drug type=str
-    drug put in the medium for analysis (can be used to add a second condition)
 - tolerance type=float
-    noise tolerance for the baseline (xstd)
-- bead type=str
-    number of the bead used for the curve
-- cell type=str
-    number of the cell used for the curve
-- couple type=str
-    couple bead number and cell number
+    noise tolerance for the baseline (as x STD of baseline)
 
-### Theoretical data present in the headers of the files
+The following one identify the experiments and can be used to pool them for stats or plotting
+    
+- Date type=str
+    date of creation of the curve file (day of the experiment)
+- Hour type=str
+    time of creation of the curve file (time of the experiment)
+    
+The following identify the conditions of the experiments
+    
+- condition type=str
+    condition applied to the analysis set (ex : antibodies decorating the bead)
+- drug type=str
+    name of drug that is used to perturb the system, if used (can be used to add a second condition, ex : surface treatment for cell attachment)
+    
+The following ones can be used for pooling the data per bead, per cell, per couple...
+
+- bead type=str
+    number of the bead used during measurements
+- cell type=str
+    number of the cell used during measurements
+- couple type=str
+    couple bead number and cell number, as an identifier
+
+### Data coming from the OT GUI and present in the headers of the files
+
[…]
