### 2023-01-23 #4877 @phpuech
Hi,
hope these words will find you well. Happy new year to all of you.
we proceeded with reviewing the article
- we edited the text
- we modified the README to make it more clear
- the last point is under work
Sincerely
PH Puech
