+++ .github/workflows/ci_pipeline.yml
@@ -8,7 +8,7 @@ on:
 
 jobs:
   Basic-CI:
-  runs-on: ${{ matrix.os }}
+    runs-on: ${{ matrix.os }}
     strategy:
       fail-fast: false
       matrix:
