### 2023-03-02 7c7a420
Update test_curve.py
