### 2023-03-28 45d051b
Update paper.bib
modified ref Manca et al
