+++ JOSS/paper.bib
@@ -87,8 +87,10 @@
 }
 
 @article{manca_membrane:2022,
-	title = {Membrane tubes pulling by optical tweezers probe the interaction of immune synapse components with actin cytoskeleton ({In preparation, tentative} title)},
+	title = {Probing mechanical interaction of immune receptors and cytoskeleton by membrane nanotube extraction},
 	author = {Manca, Fabio and Eich, Gautier and N'Dao, Omar and Normand, Lucie and Sengupta, Kheya and Limozin, Laurent and Puech, Pierre-Henri},
+	journal = {bioRxiv},
+	doi = {https://doi.org/10.1101/2022.09.15.508080},
 	year = {2022},
 }
 
