### 2023-05-04 d97c5a9
Addition of one new figure, two sections
