### 2023-05-06 #4877 @phpuech
hi @prashjha @florian-huber @ryEllison
we updated the article and repo to answer all the points that were raised, in particular by @florian-huber
- we added a figure to describe the overview of the code structure, and provided a link to the readthedoc documentation of the functions (in the article)
- we added an example of use, with the succession of possible quantification and use for the biologist (in the article)
- we documented the contribution guidelines (on the git repo)
all the best
ph puech, for the authors
