### 2023-06-26 #4877 @florian-huber
@phpuech @prashjha  Now the repo contains a contribution instruction, and the article now has an overview figure of the codes modules and an example section.

The example section needs some edits, though, :
- At some place it reads ", ...". Not sure if that means etc. or is just not yet finished
(e.g. "The user can adjust the thresholds for detecting the adhesion and potential
tubes (eg. duration and force amplitude), the threshold for optical artifacts, …")
- seems like some of this was supposed to be bullet points (which doesn't show in the pdf), so needs markdown editing.
