### 2023-06-26 08796b9
Update paper.md
modifications in the example section following reviews of florian huber : 
- markdown for bullets
- text for precising threshold uses
