### 2023-06-26 #4877 @phpuech
@florian-huber @prashjha
Hi,
thanks a lot for your comments
I corrected the text to make it more clear and modified the bullets
all the best
ph puech
