### 2023-08-17 ec62145
Update paper.md
Typos correction following Editor suggestions
