+++ JOSS/paper.md
@@ -151,9 +151,9 @@ features in the field of force measurements by optical tweezers.
 
 # Example usage
 
-The prototypical experiments consist in using optical tweezers to contact a cell with a trapped bead decorated with protein  , in order to obtain some membrane tube pulling events upon separation. The first use of the present software is to ensure that such a pulling is specific to the protein decoration used. In this case, two sets of data curves (force vs. time) are obtained using an adhesive protein and a control protein (in our case an antibody specific for a surface protein of the cell and a second antibody, isotype to the first one, but not recognizing any surface protein a priori).
+The prototypical experiments consist in using optical tweezers to contact a cell with a trapped bead decorated with protein, in order to obtain some membrane tube pulling events upon separation. The first use of the present software is to ensure that such a pulling is specific to the protein decoration used. In this case, two sets of data curves (force vs. time) are obtained using an adhesive protein and a control protein (in our case an antibody specific for a surface protein of the cell and a second antibody, isotype to the first one, but not recognizing any surface protein a priori).
 
-The software allows, after an automated pre-processing step, a inspection of the data by the user in order to amend the type selection and potentially refine, on a case per case basis, the data processing. The user can adjust several detection thresholds, eg. for determining the duration and force amplitude that characterize adhesion events and potential tubes, and for tuning the detection and correction for optical artifacts. 
+The software allows, after an automated pre-processing step, a inspection of the data by the user in order to amend the type selection and potentially refine, on a case per case basis, the data processing. The user can adjust several detection thresholds, e.g. for determining the duration and force amplitude that characterize adhesion events and potential tubes, and for tuning the detection and correction for optical artifacts. 
 
 The outcome will provide graphical representations to:
 
@@ -169,9 +169,9 @@ The outcome will provide graphical representations to:
 
 -  measure the amount of tubes in the overall population.
 
-The output file can then be reloaded using ad-hoc procedures (eg. in Python) to merge or sort the data, plot, perform statistical analysis,... on relevant biophysical parameters described in https://github.com/phpuech/OT_Analysis.
+The output file can then be reloaded using ad-hoc procedures (e.g. in Python) to merge or sort the data, plot, perform statistical analysis on relevant biophysical parameters described in https://github.com/phpuech/OT_Analysis.
 
-The results of the analysis can then be used to feedback on the experimental parameters (contact force, contact duration, molecular densities on the beads, bead size,...) and / or compare different molecules in order to examine a biological effect (for this last point, see [@manca_membrane:2022])
+The results of the analysis can then be used to feedback on the experimental parameters (contact force, contact duration, molecular densities on the beads, bead size) and / or compare different molecules in order to examine a biological effect (for this last point, see [@manca_membrane:2022]).
 
 # Functionality documentation
 
