### 2023-08-17 #4877 @phpuech
> @phpuech could you please correct the 'INVALID DOIs' error?

hi @prashjha
i edited the bib file to remove the https/
i also corrected the typos you pointed out to us

let me know if anything else is needed from us !
thanks a lot for your help
sincerely
ph puech, for the authors
