### 2023-08-17 cca6341
Update paper.bib
Corrected DOI following Editor comment
