+++ JOSS/paper.bib
@@ -90,7 +90,7 @@
 	title = {Probing mechanical interaction of immune receptors and cytoskeleton by membrane nanotube extraction},
 	author = {Manca, Fabio and Eich, Gautier and N'Dao, Omar and Normand, Lucie and Sengupta, Kheya and Limozin, Laurent and Puech, Pierre-Henri},
 	journal = {bioRxiv},
-	doi = {https://doi.org/10.1101/2022.09.15.508080},
+	doi = {10.1101/2022.09.15.508080},
 	year = {2022},
 }
 
