### 2023-09-12 #4877 @prashjha
Hi @phpuech, please confirm if the version 'v0.2.86' is correct.

Also, the title of your zenodo archive must match the title of the JOSS paper. Right now, it says 'phpuech/OT_Analysis: 0.1'. Also, the version of your code repository in the zenodo archive and in this JOSS issue should be the same. Let me know what the correct version is (if it is still 'v0.2.86', please confirm). Update the version of your code in zenodo archive as well.
