### 2023-09-14 8abdb28
Update conf.py
version 0.2
