### 2023-09-14 59eb898
Update setup.cfg
version 0.2
