+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = OT_Analysis
-version = 0.1.3
+version = 0.2.0
 author = Thierry GALLIANO (LAI)
 author_email = thierry.galliano@etu.univ-amu.fr
 
