### 2023-09-14 29ca28d
Update setup.cfg
