### 2023-09-14 #4877 @phpuech
hi @prashjha
we updated the doi to be : https://doi.org/10.5281/zenodo.8345307
the version we released is now 0.2.1
we do not understand from where the 0.2.86 was appearing on your side... could you please tell us ?
sincerely
ph puech
