### 2023-09-15 #4877 @phpuech
> Hi @phpuech, the archive title is still not correct. Please remove 'phpuech/OT_Analysis: ' so that the title matches the JOSS paper title.

hi @prashjha, i changed the name of the archive - is it ok now ? all the best
