### 2023-10-01 #4877 @arfon
@phpuech – I think the issue here is that you are using the same reference in multiple places (`fig:dessin` or similar).
