### 2023-10-01 da8bb5a
Update paper.md
change figure labels for solving compilation issues
