### 2023-10-01 #4877 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
