### 2023-10-02 #4877 @Kevin-Mattheus-Moerman
@prashjha thanks for editing here. I'll take it from here. Note you can now call `@editorialbot create post-review checklist` to generate a final checklist which will help identify some final things (e.g. the license/version/title on the archive should match those for the software/paper etc. ).
