### 2023-10-04 #4877 @phpuech
> @phpuech 👋 Could you address the points above? Thanks

hi,
i am currently away from the lab - i'll do my best to address the list asap, more likely in the coming days
i also have to modify a ref (manca et al) which was published two weeks ago
best
