### 2023-10-04 b098310
Update paper.bib
changed ref manca et al from biorxiv to published one
