+++ JOSS/paper.bib
@@ -86,12 +86,14 @@
 	pages = {210256},
 }
 
-@article{manca_membrane:2022,
+@article{manca_membrane:2023,
 	title = {Probing mechanical interaction of immune receptors and cytoskeleton by membrane nanotube extraction},
 	author = {Manca, Fabio and Eich, Gautier and N'Dao, Omar and Normand, Lucie and Sengupta, Kheya and Limozin, Laurent and Puech, Pierre-Henri},
-	journal = {bioRxiv},
-	doi = {10.1101/2022.09.15.508080},
-	year = {2022},
+	journal = {Scientific Reports},
+	doi = {10.1038/s41598-023-42599-9},
+	year = {2023},
+	pages = {15652},
+	number = {13}
 }
 
 @article{muller_nanite:2019,
