### 2023-10-04 7bfe06b
Update paper.md
changed citation to manca et al from 2022 to 2023
