### 2023-10-05 30f45ab
Update paper.md
text modifications following comments of editor
