+++ JOSS/paper.md
@@ -36,13 +36,13 @@ Optical tweezers are a light-based technique for micromanipulating
 objects. It allows to move objects such as microbeads and cells, and to
 record minute forces down to a few pN, which makes it a technique very
 well adapted to mechanical measurements on living cells
-[@gennerich_optical:2017]. We are interested in the mechano-transduction
+[@gennerich_optical:2017]. We are interested in the mechanotransduction
 properties of lymphocytes. We seek to dissect the effect of forces and
 cell mechanics on the cellular response, in the context of the immune
 system. T cell mechanotransduction has been recently demonstrated to be
 instrumental in the finesse and accuracy of the response of the latter
-@puech_mechanotransduction:2021]. Aside, cells can exert forces when
-performing their action, eg. cytotoxic T cells are using forces to kill
+@puech_mechanotransduction:2021]. In addition, cells can exert forces when
+performing their action, e.g. cytotoxic T cells are using forces to kill
 target infected cells [@basu_cytotoxic:2016].
 
 Using optical tweezers and specifically decorated beads as handles, we
@@ -52,7 +52,7 @@ tension of adherent cells [@diz-munoz_control:2010]. By varying the
 antibodies that are used to decorate the beads, we select the molecule
 type we specifically pull on, and we then explore the molecules which
 are characteristic of the immune synapse, which is one of the key
-organisational structures that have profound implications in T cell
+organizational structures that have profound implications in T cell
 recognition and action [@baldari_immune:2017].
 
 Using this approach, we probe not only the forces of recognition of the
@@ -61,7 +61,7 @@ extracellular bridges, we probe the cytosolic link of the probed
 molecule to the cytoskeleton. Such a link has been proposed to be
 instrumental in the way T cells can apply or feel forces through the
 molecule. A theoretical model has been built and will be reported in a
-dedicated article [@manca_membrane:2023]. Aside, we will demonstrate the
+dedicated article [@manca_membrane:2023]. Furthermore, we will demonstrate the
 application of the software on full data.
 
 ![Schematics of the experiment. A : Sequence of events (a) antibody
@@ -79,13 +79,13 @@ of time (among other parameters), in the three directions of space,
 obtained in large quantities (at least 10 per cell / bead couple, and up
 to 20 couples tested per sample), containing rich and detailed features
 that can relate to molecular and/or cellular mechanics that our model
-explores. It is therefore needed to standardize and semi-automatize data
+explores. It is therefore needed to standardize and semi-automate data
 analysis to help the experimentalist, often a biologist, to extract
 relevant features from the experimental data sets.
 
 # Statement of need
 
-The avalaible software that comes with the optical tweezers setup
+The available software that comes with the optical tweezers setup
 includes a data processing system, with a GUI, that allows the user to
 follow pre-defined data processing schemes. Even if it already allows
 the typical user to observe, manipulate, quantify and convert to plain
@@ -99,8 +99,8 @@ of preexisting data processing functions to perform the (time) expensive
 analysis required. This may introduce bias in the data, which may impair
 user-to-user data comparison.
 
-Aside, almost none, if any, open-source software has been proposed to
-the community eg. via GitHub or GitLab for quantifying optical tweezers
+In addition, almost none, if any, open-source software has been proposed to
+the community e.g. via GitHub or GitLab for quantifying optical tweezers
 experiments with living cells, while some have been proposed for Atomic
 Force Microscopy force mode [@muller_nanite:2019], with a modular
 capability. Among them, one can find, with the keywords \"optical
@@ -128,11 +128,11 @@ if needed.](dessin.png){#fig:dessin width="\\linewidth"}
 
[…]
