### 2023-10-05 #4877 @phpuech
hi @Kevin-Mattheus-Moerman
i modified the text accordingly to yout comments and recommendations
best
p

> @phpuech can you have a look at the above checklist? In particular the archive title should match the paper title (there is a minor difference in terms of _ or -) , the archive listed license should match your software license (currently says "other open). In addition the author set does not match, and ORCID's can be added for the archive link too. Please manually edit these aspects. In addition I have checked the paper, and the the remaining points below that require your attention:
>
>     * [ ]  Currently the paper features `mechano-transduction` and `mechanotransduction`, please choose one convention and use it throughout the paper.
>
>     * [ ]  Currently both Brittish (e.g. `organisational`) and American English (e.g. `characterize`, `standardize` ) words are used. Words like
>
>     * [ ]  `semi-automatize` should likely be `semi-automate`
>
>     * [ ]  Check `avalaible` which should read `available`
>
>     * [ ]  Check `eg.` which should read `e.g.`
>
>     * [ ]  Check `mis-alignments` which could be written `misalignments`
>
>     * [ ]  Check `aknowledge` which should read `acknowledge`
>
>     * [ ]  Recommendation: You may consider using italic text for Latin text like _a priori_.
>
>     * [ ]  Recommendation: Although `semi-automatize` may be correct, perhaps `semi-automate` is more common.
>
>     * [ ]  Comment: The authors use `Aside...` instead of `In addition...` or `Furthermore`, I would argue the latter two sound more natural (and that `aside` feels similar to `besides` rather than `in addition`), but this may be a personal preference. Perhaps check if your use of `aside` is correct and decide to keep it or rephrase.
