### 2023-10-07 adf0d8e
Update paper.md
modified text when revising to have proper tense (citation of manca et al)
