### 2023-10-16 #4877 @Kevin-Mattheus-Moerman
@phpuech congratulations on this publication in JOSS!

@prashjha thanks for editing!

And a special thanks to the reviewers: @ryEllison, @florian-huber
